refactor : 쿠폰 종류 제한되도록 수정
