--- src/main/kotlin/com/example/estdelivery/domain/coupon/FixDiscountCoupon.kt
@@ -1,10 +1,0 @@
-package com.example.estdelivery.domain.coupon
-
-class FixDiscountCoupon(
-    val discountAmount: Int,
-    override val name: String,
-    override val description: String,
-    private val couponType: CouponType
-) : Coupon(name, description, couponType) {
-
-}
