+++ src/test/kotlin/com/example/estdelivery/domain/coupon/CouponTest.kt
@@ -4,8 +4,8 @@ import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 
 class CouponTest : FreeSpec({
-    val 나눠준_비율_할인_쿠폰 = RateDiscountCoupon(10, "10% 할인 쿠폰", "10% 할인 쿠폰 설명", CouponType.IS_HAND_OUT)
-    val 게시된_고정_할인_쿠폰 = FixDiscountCoupon(1000, "1000원 할인 쿠폰", "1000원 할인 쿠폰 설명", CouponType.IS_PUBLISHED)
+    val 나눠준_비율_할인_쿠폰 = Coupon.RateDiscountCoupon(10, "10% 할인 쿠폰", "10% 할인 쿠폰 설명", CouponType.IS_HAND_OUT)
+    val 게시된_고정_할인_쿠폰 = Coupon.FixDiscountCoupon(1000, "1000원 할인 쿠폰", "1000원 할인 쿠폰 설명", CouponType.IS_PUBLISHED)
 
     "게시된 쿠폰인지 확인 할 수 있다." {
         게시된_고정_할인_쿠폰.isPublished() shouldBe true
